requirements.txt
torch
torchvision
pycocotools
matplotlib
tqdm
Pillow